requirements.txt
streamlit
pandas
matplotlib
google-analytics-data
google-auth
streamlit-cookies-manager
protobuf